pyproject.toml
[tool.poetry]
name = "all-hail-chaos"
version = "2.0.0"
description = "My awesome Discord bot"
authors = ["Waylander <62927863+Faholan@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
toml = "^0.10.2"
psycopg = {extras = ["binary", "pool"], version = "^3.0.16"}
lavalink = "^3.1.7"
humanize = "^4.3.0"
psutil = "^5.9.1"
"discord.py" = {extras = ["voice", "speed"], version = "^2.0.0"}
ujson = "^5.4.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
isort = "^5.10.1"
pylama = "^8.4.1"
pydocstyle = "^6.1.1"
mypy = "^0.971"
flake8 = "^5.0.4"
bandit = "^1.7.4"
pylint = "^2.14.5"
pycodestyle = "^2.9.1"
types-toml = "^0.10.8"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
line-length = 88
